pyproject.toml
[project]
name = "hw-snowflake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dbt-core>=1.11.6",
    "dbt-duckdb>=1.10.1",
    "dbt-snowflake>=1.11.2",
    "dotenv>=0.9.9",
    "duckdb>=1.4.4",
    "pandas>=3.0.0",
    "pyarrow>=23.0.0",
    "snowflake-connector-python>=4.2.0",
]
